+++ common/src/main/java/me/rufia/fightorflight/entity/ai/tasks/FOFDefendSelfTask.java
@@ -14,23 +14,27 @@ public class FOFDefendSelfTask {
                                 context.present(MemoryModuleType.HURT_BY_ENTITY),
                                 context.registered(MemoryModuleType.ATTACK_TARGET)
                         )
-                        .apply(context, (hurtByAccessor, attackTargetAccessor) -> ((serverLevel, pokemonEntity, l) -> {
-                            var hurtByEntity = context.get(hurtByAccessor);
-                            if (hurtByEntity instanceof PokemonEntity pokemonEntity1) {
-                                if (PokemonUtils.tryToAvoidWildShiny(pokemonEntity1)) {
+                        .apply(context, (hurtByAccessor, attackTargetAccessor) -> ((serverLevel, livingEntity, l) -> {
+                            if (livingEntity instanceof PokemonEntity pokemonEntity) {
+                                if (PokemonUtils.shouldAvoid(pokemonEntity)) {
                                     return false;
                                 }
+                                var hurtByEntity = context.get(hurtByAccessor);
+                                if (hurtByEntity instanceof PokemonEntity pokemonEntity1) {
+                                    if (PokemonUtils.tryToAvoidWildShiny(pokemonEntity1)) {
+                                        return false;
+                                    }
+                                }
+                                var targetOpt = pokemonEntity.getBrain().getMemory(MemoryModuleType.ATTACK_TARGET);
+                                if (targetOpt.isPresent()) {
+                                    LivingEntity target = targetOpt.get();
+                                    //Try to add a priority queue here, or at least allow the Pokemon to attack the entity that hurts it.
+                                } else {
+                                    attackTargetAccessor.set(hurtByEntity);
+                                }
+                                return true;
                             }
-                            var targetOpt = pokemonEntity.getBrain().getMemory(MemoryModuleType.ATTACK_TARGET);
-                            if (targetOpt.isPresent()) {
-                                LivingEntity target = targetOpt.get();
-                                //Try to add a priority queue here, or at least allow the Pokemon to attack the entity that hurts it.
-                            } else {
-                                attackTargetAccessor.set(hurtByEntity);
-                            }
-
-
-                            return true;
+                            return false;
                         })));
     }
 }